Modularize Sound Pack Manager dialog (UI/state/mappings/ops) and shrink dialog file (#37)
* SoundPackManagerDialog: extract _load_pack_meta/_save_pack_meta; refactor simple mapping and removal to use helpers. Tests: 561 passed.

* SoundPackManagerDialog: use pack.json helpers in _on_browse_mapping_file. Tests: 561 passed.

* SoundPackManagerDialog: use pack.json helpers in edit metadata flow. Tests: 561 passed.

* SoundPackManagerDialog: finish converting remaining pack.json writes to helper/try blocks; no behavior change. Tests: 561 passed.

* SoundPackManagerDialog: extract import/create/duplicate/delete operations to soundpack_manager/ops.py and delegate. Tests: 561 passed.

* SoundPackManagerDialog: extract mapping handlers to soundpack_manager/mappings.py and delegate; continue shrinking monolith. Tests: 561 passed.

* SoundPackManagerDialog: move AlertCategoryItem + FRIENDLY_ALERT_CATEGORIES into soundpack_manager/constants and import. Shrink dialog file further. Tests passing.

* SoundPackManagerDialog: extract UI creation (dialog, panels, buttons) to soundpack_manager/ui.py and delegate. Dialog file now 662 lines. All tests passing.

* SoundPackManagerDialog: extract state/update helpers (load/refresh/update) to soundpack_manager/state.py; delegate; dialog now ~500 lines. All tests passing (561).

* SoundPackManagerDialog: move preview handlers to mappings module and delegate; keep meta load/save via ops. Ruff + tests pass (561).
