+++ src/accessiweather/dialogs/soundpack_manager_dialog.py
@@ -6,12 +6,7 @@ in Settings > General and is not changed from this manager.
 """
 
 import asyncio
-import contextlib
-import json
 import logging
-import shutil
-import zipfile
-from dataclasses import dataclass
 from pathlib import Path
 
 import toga
@@ -30,44 +25,8 @@ from ..notifications.sound_player import (  # noqa: F401
 # Community packs imports
 from ..services.community_soundpack_service import CommunitySoundPackService
 
-
-@dataclass
-class AlertCategoryItem:
-    """Data class for alert category selection items."""
-
-    display_name: str
-    technical_key: str
-
-    def __str__(self) -> str:  # Shown by Selection and debug; only show friendly name
-        return self.display_name
-
-    def __repr__(self) -> str:  # Prevent noisy reprs in lists/debug UIs
-        return f"{self.display_name}"
-
-
-# Friendly alert categories for mapping (display name, technical key)
-# Updated to match canonical keys from alert_sound_mapper
-FRIENDLY_ALERT_CATEGORIES = [
-    ("Tornado Warnings", "tornado_warning"),
-    ("Flood Warnings", "flood_warning"),
-    ("Heat Advisories", "heat_advisory"),
-    ("Thunderstorm Warnings", "thunderstorm_warning"),
-    ("Winter Storm Warnings", "winter_storm_warning"),
-    ("Hurricane Warnings", "hurricane_warning"),
-    ("Wind Warnings", "wind_warning"),
-    ("Fire Weather Warnings", "fire_warning"),
-    ("Air Quality Alerts", "air_quality_alert"),
-    ("Fog Advisories", "fog_advisory"),
-    ("Ice Warnings", "ice_warning"),
-    ("Snow Warnings", "snow_warning"),
-    ("Dust Warnings", "dust_warning"),
-    ("Generic Warning", "warning"),
-    ("Generic Watch", "watch"),
-    ("Generic Advisory", "advisory"),
-    ("Generic Statement", "statement"),
-    ("General Alert", "alert"),
-    ("General Notification", "notify"),
-]
+# Import shared constants/datatypes for categories
+from .soundpack_manager.constants import FRIENDLY_ALERT_CATEGORIES
 
 logger = logging.getLogger(__name__)
 
@@ -110,293 +69,30 @@ class SoundPackManagerDialog:
         self._create_dialog()
 
     def _load_sound_packs(self) -> None:
-        """Load all available sound packs."""
-        self.sound_packs.clear()
+        from .soundpack_manager.state import load_sound_packs as _load
 
-        if not self.soundpacks_dir.exists():
-            return
-
-        for pack_dir in self.soundpacks_dir.iterdir():
-            if not pack_dir.is_dir():
-                continue
-
-            pack_json = pack_dir / "pack.json"
-            if not pack_json.exists():
-                continue
-
-            try:
-                with open(pack_json, encoding="utf-8") as f:
-                    pack_data = json.load(f)
-
-                pack_data["directory"] = pack_dir.name
-                pack_data["path"] = pack_dir
-                self.sound_packs[pack_dir.name] = pack_data
-
-            except Exception as e:
-                logger.error(f"Failed to load sound pack {pack_dir.name}: {e}")
+        _load(self)
 
     def _create_dialog(self) -> None:
         """Create the sound pack manager dialog."""
-        self.dialog = toga.Window(
-            title="Sound Pack Manager",
-            size=(800, 600),
-            resizable=True,
-        )
+        from .soundpack_manager.ui import create_dialog_ui as _ui
 
-        main_box = toga.Box(style=Pack(direction=COLUMN, margin=10, flex=1))
-
-        # Title
-        title_label = toga.Label(
-            "Sound Pack Manager", style=Pack(font_size=16, font_weight="bold", margin_bottom=10)
-        )
-        main_box.add(title_label)
-
-        # Content area
-        content_box = toga.Box(style=Pack(direction=ROW, flex=1, margin_bottom=10))
-
-        # Left panel - Sound pack list
-        left_panel = self._create_pack_list_panel()
-        content_box.add(left_panel)
-
-        # Right panel - Pack details and sounds
-        right_panel = self._create_pack_details_panel()
-        content_box.add(right_panel)
-
-        main_box.add(content_box)
-
-        # Bottom buttons
-        button_box = self._create_button_panel()
-        main_box.add(button_box)
-
-        self.dialog.content = main_box
-
-        # Populate the pack list with loaded sound packs
-        self._refresh_pack_list()
-
-        # Select current pack if available (for focus), but do not change app setting here
-        if self.current_pack in self.sound_packs:
-            self.selected_pack = self.current_pack
-            # Find and set the current pack in the detailed list
-            for item in self.pack_list.data:
-                if item.pack_id == self.current_pack:
-                    # Note: DetailedList selection is read-only, so we can't set it programmatically
-                    # Just update the pack details instead
-                    break
-            self._update_pack_details()
+        _ui(self)
 
     def _create_pack_list_panel(self) -> toga.Box:
-        """Create the sound pack list panel."""
-        panel = toga.Box(style=Pack(direction=COLUMN, flex=1, margin_right=10))
+        from .soundpack_manager.ui import create_pack_list_panel as _panel
 
-        # Panel title
-        title_label = toga.Label(
-            "Available Sound Packs", style=Pack(font_weight="bold", margin_bottom=5)
-        )
-        panel.add(title_label)
-
-        # Pack list
-        self.pack_list = toga.DetailedList(
-            on_select=self._on_pack_selected,
-            style=Pack(flex=1, margin_bottom=10),
-        )
-        panel.add(self.pack_list)
-
-        # Import button
-        self.import_button = toga.Button(
-            "Import Sound Pack", on_press=self._on_import_pack, style=Pack(width=150)
-        )
-        panel.add(self.import_button)
-
-        # Quick hint: Active pack is selected in Settings > General
-        panel.add(
-            toga.Label(
-                "Hint: Select your active pack in Settings > General.",
-                style=Pack(margin_top=8, font_style="italic"),
-            )
-        )
-
-        return panel
+        return _panel(self)
 
     def _create_pack_details_panel(self) -> toga.Box:
-        """Create the pack details panel."""
-        panel = toga.Box(style=Pack(direction=COLUMN, flex=2, margin_left=10))
+        from .soundpack_manager.ui import create_pack_details_panel as _panel
 
-        # Panel title
-        title_label = toga.Label(
-            "Sound Pack Details", style=Pack(font_weight="bold", margin_bottom=5)
-        )
-        panel.add(title_label)
-
-        # Pack info box
-        self.pack_info_box = toga.Box(
-            style=Pack(direction=COLUMN, margin=10, background_color="#f0f0f0")
-        )
-
-        self.pack_name_label = toga.Label(
-            "No pack selected", style=Pack(font_size=14, font_weight="bold", margin_bottom=5)
-        )
-        self.pack_info_box.add(self.pack_name_label)
-
-        self.pack_author_label = toga.Label("", style=Pack(margin_bottom=5))
-        self.pack_info_box.add(self.pack_author_label)
-
-        self.pack_description_label = toga.Label("", style=Pack(margin_bottom=10))
-        self.pack_info_box.add(self.pack_description_label)
-
-        panel.add(self.pack_info_box)
-
-        # Sounds section
-        sounds_label = toga.Label(
-            "Sounds in this pack:", style=Pack(font_weight="bold", margin=(10, 0, 5, 0))
-        )
-        panel.add(sounds_label)
-
-        # Sound selection
-        self.sound_selection = toga.Selection(
-            items=[],
-            accessor="display_name",
-            on_change=self._on_sound_selected,
-            style=Pack(flex=1, margin_bottom=10),
-        )
-        panel.add(self.sound_selection)
-
-        # Mapping controls header and help text (no widget tooltips; Toga tooltips are for Commands)
-        mapping_header = toga.Box(style=Pack(direction=ROW, margin_bottom=5))
-        # WCAG: Keep labels concise; provide verbose guidance in separate help text.
-        mapping_label = toga.Label(
-            "Alert category:",
-            style=Pack(font_weight="bold", margin_right=5),
-        )
-        mapping_header.add(mapping_label)
-        panel.add(mapping_header)
-        # Mapping controls: key dropdown + file picker + preview
-        mapping_row = toga.Box(style=Pack(direction=ROW, margin_bottom=10))
-        self.mapping_key_selection = toga.Selection(
-            items=[
-                AlertCategoryItem(display_name=name, technical_key=key)
-                for name, key in FRIENDLY_ALERT_CATEGORIES
-            ],
-            accessor="display_name",
-            on_change=self._on_mapping_key_change,
-            style=Pack(width=260, margin_right=10),
-        )
-        # Accessibility: keep label concise; attach verbose help as description of the control
-        with contextlib.suppress(Exception):
-            self.mapping_key_selection.aria_label = "Alert category"
-            self.mapping_key_selection.aria_description = (
-                "Select from common weather alert categories. Each category maps to technical keys used by weather services. "
-                "Use the custom mapping field below for specific alert types not listed."
-            )
-        self.mapping_file_input = toga.TextInput(
-            readonly=True, placeholder="Select audio file...", style=Pack(flex=1, margin_right=10)
-        )
-        self.mapping_browse_button = toga.Button(
-            "Browse...", on_press=self._on_browse_mapping_file, style=Pack(margin_right=10)
-        )
-        self.mapping_preview_button = toga.Button(
-            "Preview", on_press=self._on_preview_mapping, enabled=False
-        )
-        mapping_row.add(self.mapping_key_selection)
-        mapping_row.add(self.mapping_file_input)
-        mapping_row.add(self.mapping_browse_button)
-        mapping_row.add(self.mapping_preview_button)
-        panel.add(mapping_row)
-        # Simpler controls for non-technical users: Add a custom key mapping directly
-        simple_map_box = toga.Box(style=Pack(direction=ROW, margin_bottom=10))
-        simple_label = toga.Label("Add or change a mapping:", style=Pack(margin_right=10))
-        self.simple_key_input = toga.TextInput(
-            placeholder="e.g., excessive_heat_warning or tornado_warning",
-            style=Pack(width=260, margin_right=10),
-        )
-        self.simple_file_button = toga.Button(
-            "Choose Sound...", on_press=self._on_simple_choose_file, style=Pack(margin_right=10)
-        )
-        self.simple_remove_button = toga.Button(
-            "Remove Mapping", on_press=self._on_simple_remove_mapping
-        )
-        simple_map_box.add(simple_label)
-        simple_map_box.add(self.simple_key_input)
-        simple_map_box.add(self.simple_file_button)
-        simple_map_box.add(self.simple_remove_button)
-        panel.add(simple_map_box)
-
-        # Sound action buttons removed; preview is available per alert category mapping
-        return panel
+        return _panel(self)
 
     def _create_button_panel(self) -> toga.Box:
-        """Create the bottom button panel."""
-        button_box = toga.Box(style=Pack(direction=ROW))
+        from .soundpack_manager.ui import create_button_panel as _btns
 
-        # Add flexible space to push buttons to the right
-        button_box.add(toga.Box(style=Pack(flex=1)))
-
-        # Create New Pack
-        self.create_button = toga.Button(
-            "Create New",
-            on_press=self._on_create_pack,
-            style=Pack(margin_right=10),
-        )
-        button_box.add(self.create_button)
-
-        # Create with Wizard (guided)
-        self.create_wizard_button = toga.Button(
-            "Create with Wizard",
-            on_press=self._on_create_pack_wizard,
-            style=Pack(margin_right=10),
-        )
-        button_box.add(self.create_wizard_button)
-
-        # Browse Community Packs
-        self.browse_community_button = toga.Button(
-            "Browse Community",
-            on_press=self._on_browse_community_packs,
-            style=Pack(margin_right=10),
-            enabled=self.community_service is not None,
-        )
-        button_box.add(self.browse_community_button)
-
-        # Duplicate Selected Pack
-        self.duplicate_button = toga.Button(
-            "Duplicate",
-            on_press=self._on_duplicate_pack,
-            enabled=False,
-            style=Pack(margin_right=10),
-        )
-        button_box.add(self.duplicate_button)
-
-        # Edit Metadata of Selected Pack
-        self.edit_button = toga.Button(
-            "Edit",
-            on_press=self._on_edit_pack,
-            enabled=False,
-            style=Pack(margin_right=10),
-        )
-        button_box.add(self.edit_button)
-
-        # Open Selected Pack (focus/manage only; does not change active pack)
-        self.select_button = toga.Button(
-            "Open",
-            on_press=self._on_open_pack,
-            enabled=False,
-            style=Pack(margin_right=10, background_color="#4CAF50", color="#ffffff"),
-        )
-        button_box.add(self.select_button)
-
-        # Delete Selected Pack
-        self.delete_button = toga.Button(
-            "Delete Pack",
-            on_press=self._on_delete_pack,
-            enabled=False,
-            style=Pack(margin_right=10, background_color="#ff4444", color="#ffffff"),
-        )
-        button_box.add(self.delete_button)
-
-        self.close_button = toga.Button(
-            "Close", on_press=self._on_close, style=Pack(margin_right=0)
-        )
-        button_box.add(self.close_button)
-
-        return button_box
+        return _btns(self)
 
     def _on_pack_selected(self, widget) -> None:
         """Handle pack selection."""
@@ -418,178 +114,23 @@ class SoundPackManagerDialog:
         )  # Don't allow deleting default pack
 
     def _update_pack_details(self) -> None:
-        """Update the pack details display."""
-        if not self.selected_pack or self.selected_pack not in self.sound_packs:
-            self.pack_name_label.text = "No pack selected"
-            self.pack_author_label.text = ""
-            self.pack_description_label.text = ""
-            self.sound_selection.items = []
-            return
+        from .soundpack_manager.state import update_pack_details as _upd
 
-        pack_info = self.sound_packs[self.selected_pack]
-
-        # Update pack info
-        self.pack_name_label.text = pack_info.get("name", self.selected_pack)
-        self.pack_author_label.text = f"Author: {pack_info.get('author', 'Unknown')}"
-        self.pack_description_label.text = pack_info.get("description", "No description available")
-
-        # Update sound selection
-        sounds = pack_info.get("sounds", {})
-        sound_items = []
-        for sound_name, sound_file in sounds.items():
-            sound_path = pack_info["path"] / sound_file
-            status = "✓" if sound_path.exists() else "✗"
-
-            # Use friendly name if available for recognized keys
-            friendly_name = sound_name.title()
-            for display_name, technical_key in FRIENDLY_ALERT_CATEGORIES:
-                if technical_key == sound_name:
-                    friendly_name = display_name
-                    break
-
-            display_name = f"{friendly_name} ({sound_file}) - {status} {'Available' if sound_path.exists() else 'Missing'}"
-            sound_items.append(
-                {
-                    "display_name": display_name,
-                    "sound_name": sound_name,
-                    "sound_file": sound_file,
-                    "exists": sound_path.exists(),
-                }
-            )
-
-        self.sound_selection.items = sound_items
-
-        # Update mapping selection to a friendly category that has a mapping in this pack
-        try:
-            if self.mapping_key_selection is not None:
-                sounds_keys = set(sounds.keys())
-                mapping_items = getattr(self.mapping_key_selection, "items", []) or []
-                selected_item = None
-                for item in mapping_items:
-                    # Derive a technical key from the item, handling Toga Row wrappers
-                    technical_key = None
-                    try:
-                        # Direct attribute on dataclass
-                        technical_key = getattr(item, "technical_key", None)
-                        # Dict-style fallback
-                        if not technical_key and isinstance(item, dict):
-                            technical_key = item.get("technical_key")
-                        if not technical_key:
-                            # Toga Row case: display_name may be an AlertCategoryItem instance
-                            dn = getattr(item, "display_name", None)
-                            # If dn is a dataclass instance
-                            if dn is not None:
-                                tk2 = getattr(dn, "technical_key", None)
-                                if tk2:
-                                    technical_key = tk2
-                                else:
-                                    # If dn is a dataclass or string, try to map by friendly name
-                                    name = (
-                                        getattr(dn, "display_name", None)
-                                        if dn is not None
-                                        else None
-                                    )
-                                    if not name and isinstance(dn, str):
-                                        name = dn
-                                    if name:
-                                        for _name, _key in FRIENDLY_ALERT_CATEGORIES:
-                                            if _name == name:
-                                                technical_key = _key
-                                                break
-                    except Exception:
-                        technical_key = None
-                    if technical_key and technical_key in sounds_keys:
-                        selected_item = item
-                        break
-                if selected_item is None and mapping_items:
-                    selected_item = mapping_items[0]
-                if selected_item is not None:
-                    self.mapping_key_selection.value = selected_item
-                    # Populate file input/preview state for initial selection
-                    self._on_mapping_key_change(self.mapping_key_selection)
-        except Exception as e:
-            logger.warning(f"Failed to update mapping selection: {e}")
+        _upd(self)
 
     def _on_mapping_key_change(self, widget) -> None:
         """When the mapping key changes, populate the file input with current mapping if present."""
-        try:
-            if not self.selected_pack or self.selected_pack not in self.sound_packs:
-                return
-            pack_info = self.sound_packs[self.selected_pack]
-            sounds = pack_info.get("sounds", {})
-            key = self._get_technical_key_from_selection()
-            current = sounds.get(key, "")
-            self.mapping_file_input.value = current
-            # Enable preview if file exists
-            if current:
-                sound_path = pack_info["path"] / current
-                self.mapping_preview_button.enabled = (
-                    sound_path.exists() and sound_path.stat().st_size > 0
-                )
-            else:
-                self.mapping_preview_button.enabled = False
-        except Exception as e:
-            logger.warning(f"Failed to update mapping input: {e}")
+        from .soundpack_manager.mappings import on_mapping_key_change as _mkc
+
+        _mkc(self, widget)
 
     def _on_browse_mapping_file(self, widget) -> None:
         """Open a file dialog to choose an audio file for the selected key."""
         if not self.selected_pack or self.selected_pack not in self.sound_packs:
             return
-        display_name = self._get_display_name_from_selection()
-        technical_key = self._get_technical_key_from_selection()
-        if not technical_key:
-            return
-        try:
-            # Choose audio file
-            def _apply_file_choice(_, path=None):
-                if not path:
-                    return
-                try:
-                    pack_info = self.sound_packs[self.selected_pack]
-                    rel_name = Path(path).name  # store filename relative to pack
-                    # Update pack.json
-                    pack_json_path = pack_info["path"] / "pack.json"
-                    with open(pack_json_path, encoding="utf-8") as f:
-                        meta = json.load(f)
-                    sounds = meta.get("sounds", {})
-                    if not isinstance(sounds, dict):
-                        sounds = {}
-                    key = technical_key
-                    sounds[key] = rel_name
-                    meta["sounds"] = sounds
-                    with open(pack_json_path, "w", encoding="utf-8") as f:
-                        json.dump(meta, f, indent=2)
-                    # Copy file into pack if not already there
-                    dst = pack_info["path"] / rel_name
-                    if Path(path) != dst:
-                        try:
-                            shutil.copy2(path, dst)
-                        except Exception as copy_err:
-                            logger.warning(f"Could not copy audio file: {copy_err}")
-                    # Update in-memory data and UI
-                    pack_info["sounds"] = sounds
-                    self.mapping_file_input.value = rel_name
-                    self._update_pack_details()
-                    self.app.main_window.info_dialog(
-                        "Mapping Updated",
-                        f"Mapped '{display_name or key}' to '{rel_name}' in pack '{pack_info.get('name', self.selected_pack)}'.",
-                    )
-                except Exception as e:
-                    logger.error(f"Failed to update mapping: {e}")
-                    self.app.main_window.error_dialog(
-                        "Mapping Error", f"Failed to update mapping: {e}"
-                    )
+        from .soundpack_manager.mappings import on_browse_mapping_file as _b
 
-            self.app.main_window.open_file_dialog(
-                title="Select Audio File",
-                file_types=["wav", "mp3", "ogg", "flac"],
-                on_result=_apply_file_choice,
-            )
-        except Exception as e:
-            logger.error(f"Failed to open file dialog: {e}")
-            self.app.main_window.error_dialog(
-                "File Dialog Error", f"Failed to open file dialog: {e}"
-            )
+        _b(self, widget)
 
     def _get_selected_mapping_item(self):
         try:
@@ -598,85 +139,19 @@ class SoundPackManagerDialog:
             return None
 
     def _get_technical_key_from_selection(self) -> str | None:
-        """Safely extract the technical key from the current mapping selection."""
-        try:
-            if not self.mapping_key_selection or not self.mapping_key_selection.value:
-                return None
-            sel = self.mapping_key_selection.value
+        from .soundpack_manager.mappings import get_technical_key_from_selection as _tk
 
-            # Direct attribute (dataclass instance assigned directly)
-            tk = getattr(sel, "technical_key", None)
-            if tk:
-                return tk
-
-            # Dict-style fallback
-            if isinstance(sel, dict):
-                tk = sel.get("technical_key")
-                if tk:
-                    return tk
-
-            # Toga Row wrapper: display_name may actually be the AlertCategoryItem instance
-            dn = getattr(sel, "display_name", None)
-            if dn is not None:
-                tk2 = getattr(dn, "technical_key", None)
-                if tk2:
-                    return tk2
-                # Or try mapping from friendly name
-                name = getattr(dn, "display_name", None) if dn is not None else None
-                if not name and isinstance(dn, str):
-                    name = dn
-                if name:
-                    for friendly, key in FRIENDLY_ALERT_CATEGORIES:
-                        if friendly == name:
-                            return key
-
-            # Last resort: if selection is a plain string, normalize underscores
-            if isinstance(sel, str):
-                return sel.strip().lower().replace(" ", "_")
-
-            return None
-        except Exception:
-            return None
+        return _tk(self)
 
     def _get_display_name_from_selection(self) -> str | None:
-        try:
-            if not self.mapping_key_selection or not self.mapping_key_selection.value:
-                return None
-            sel = self.mapping_key_selection.value
-            # With AlertCategoryItem dataclass, we can directly access the attribute
-            if hasattr(sel, "display_name"):
-                return sel.display_name
-            # Fallback for backward compatibility
-            if isinstance(sel, dict):
-                return sel.get("display_name")
-            return str(sel) if sel is not None else None
-        except Exception:
-            return None
+        from .soundpack_manager.mappings import get_display_name_from_selection as _dn
+
+        return _dn(self)
 
     def _on_preview_mapping(self, widget) -> None:
-        """Preview the audio currently mapped for the selected key."""
-        try:
-            if not self.selected_pack or self.selected_pack not in self.sound_packs:
-                return
-            technical_key = self._get_technical_key_from_selection()
-            if not technical_key:
-                return
-            pack_info = self.sound_packs[self.selected_pack]
-            sounds = pack_info.get("sounds", {})
-            filename = sounds.get(technical_key) or f"{technical_key}.wav"
-            sound_path = pack_info["path"] / filename
-            if not sound_path.exists():
-                self.app.main_window.info_dialog(
-                    "Sound Not Available",
-                    f"The sound file '{filename}' is not available in this pack.",
-                )
-                return
-            from ..notifications.sound_player import play_sound_file
+        from .soundpack_manager.mappings import preview_mapping as _prev
 
-            play_sound_file(sound_path)
-        except Exception as e:
-            logger.error(f"Failed to preview mapping: {e}")
-            self.app.main_window.error_dialog("Preview Error", f"Failed to preview mapping: {e}")
+        _prev(self, widget)
 
     def _on_sound_selected(self, widget) -> None:
         """Handle sound selection (no separate preview button)."""
@@ -685,45 +160,17 @@ class SoundPackManagerDialog:
         # Preview is provided via the alert category mapping controls.
 
     def _on_preview_sound(self, widget) -> None:
-        """Preview the selected sound."""
-        if not self.sound_selection.value or not self.selected_pack:
-            return
+        from .soundpack_manager.mappings import preview_selected_sound as _ps
 
-        try:
-            sound_item = self.sound_selection.value
-            sound_name = sound_item.sound_name
-
-            # Check if the sound file exists
-            pack_info = self.sound_packs[self.selected_pack]
-            sound_path = pack_info["path"] / sound_item.sound_file
-
-            if not sound_path.exists():
-                self.app.main_window.info_dialog(
-                    "Sound Not Available",
-                    f"The sound file '{sound_item.sound_file}' is not available. This may be a placeholder sound pack.",
-                )
-                return
-
-            from ..notifications.sound_player import play_notification_sound
-
-            play_notification_sound(sound_name, self.selected_pack)
-
-        except Exception as e:
-            logger.error(f"Failed to preview sound: {e}")
-            self.app.main_window.error_dialog("Preview Error", f"Failed to preview sound: {e}")
+        _ps(self, widget)
 
     def _on_import_pack(self, widget) -> None:
         """Import a new sound pack."""
-        try:
-            # Open file dialog to select sound pack zip file
-            self.app.main_window.open_file_dialog(
-                title="Select Sound Pack ZIP File",
-                file_types=["zip"],
-                on_result=self._import_pack_file,
-            )
-        except Exception as e:
-            logger.error(f"Failed to open import dialog: {e}")
-            self.app.main_window.error_dialog("Import Error", f"Failed to open import dialog: {e}")
+        self.app.main_window.open_file_dialog(
+            title="Select Sound Pack ZIP File",
+            file_types=["zip"],
+            on_result=lambda w, path=None: self._import_pack_file(w, path),
+        )
 
     def _on_simple_choose_file(self, widget) -> None:
         """Simplified: pick a file and map it to the entered key, copying into the pack."""
@@ -734,284 +181,60 @@ class SoundPackManagerDialog:
             self.app.main_window.info_dialog("Missing Key", "Please enter a mapping key first.")
             return
 
-        def _apply(_, path=None):
-            if not path:
-                return
-            try:
-                pack_info = self.sound_packs[self.selected_pack]
-                pack_json_path = pack_info["path"] / "pack.json"
-                with open(pack_json_path, encoding="utf-8") as f:
-                    meta = json.load(f)
-                sounds = meta.get("sounds", {})
-                if not isinstance(sounds, dict):
-                    sounds = {}
-                rel_name = Path(path).name
-                sounds[key] = rel_name
-                meta["sounds"] = sounds
-                with open(pack_json_path, "w", encoding="utf-8") as f:
-                    json.dump(meta, f, indent=2)
-                dst = pack_info["path"] / rel_name
-                if Path(path) != dst:
-                    with contextlib.suppress(Exception):
-                        shutil.copy2(path, dst)
-                pack_info["sounds"] = sounds
-                self._update_pack_details()
-                self.app.main_window.info_dialog(
-                    "Mapping Updated", f"Mapped '{key}' to '{rel_name}'."
-                )
-            except Exception as e:
-                logger.error(f"Failed to update mapping: {e}")
-                self.app.main_window.error_dialog("Mapping Error", f"Failed to update mapping: {e}")
-
         self.app.main_window.open_file_dialog(
             title="Select Audio File",
             file_types=["wav", "mp3", "ogg", "flac"],
-            on_result=_apply,
+            on_result=lambda _, path=None: self._apply_simple_mapping(key, path),
         )
 
+    def _apply_simple_mapping(self, key: str, path: str | None) -> None:
+        from .soundpack_manager.mappings import apply_simple_mapping as _asm
+
+        _asm(self, key, path)
+
+    def _load_pack_meta(self, pack_info: dict) -> dict:
+        from .soundpack_manager.ops import load_pack_meta as _load
+
+        return _load(pack_info)
+
+    def _save_pack_meta(self, pack_info: dict, meta: dict) -> None:
+        from .soundpack_manager.ops import save_pack_meta as _save
+
+        _save(pack_info, meta)
+
     def _on_simple_remove_mapping(self, widget) -> None:
-        """Remove the entered mapping key from the pack.json (no file deletes)."""
-        if not self.selected_pack or self.selected_pack not in self.sound_packs:
-            return
-        key = (self.simple_key_input.value or "").strip().lower()
-        if not key:
-            self.app.main_window.info_dialog("Missing Key", "Please enter a mapping key to remove.")
-            return
-        try:
-            pack_info = self.sound_packs[self.selected_pack]
-            pack_json_path = pack_info["path"] / "pack.json"
-            with open(pack_json_path, encoding="utf-8") as f:
-                meta = json.load(f)
-            sounds = meta.get("sounds", {}) or {}
-            if key in sounds:
-                sounds.pop(key, None)
-                meta["sounds"] = sounds
-                with open(pack_json_path, "w", encoding="utf-8") as f:
-                    json.dump(meta, f, indent=2)
-                pack_info["sounds"] = sounds
-                self._update_pack_details()
-                self.app.main_window.info_dialog("Mapping Removed", f"Removed mapping for '{key}'.")
-            else:
-                self.app.main_window.info_dialog("Not Found", f"No mapping exists for '{key}'.")
-        except Exception as e:
-            logger.error(f"Failed to remove mapping: {e}")
-            self.app.main_window.error_dialog("Remove Error", f"Failed to remove mapping: {e}")
+        from .soundpack_manager.mappings import on_simple_remove_mapping as _srm
+
+        _srm(self, widget)
 
     def _import_pack_file(self, widget, path: str | None = None) -> None:
         """Import a sound pack from a ZIP file."""
-        if not path:
-            return
+        from .soundpack_manager.ops import import_pack_file as _import
 
-        try:
-            with zipfile.ZipFile(path, "r") as zip_file:
-                # Check if pack.json exists in the zip
-                if "pack.json" not in zip_file.namelist():
-                    self.app.main_window.error_dialog(
-                        "Invalid Sound Pack",
-                        "The selected file is not a valid sound pack. Missing pack.json file.",
-                    )
-                    return
-
-                # Read pack.json to get pack info
-                with zip_file.open("pack.json") as f:
-                    pack_info = json.load(f)
-
-                pack_name = pack_info.get("name", "Unknown Pack")
-                pack_id = pack_name.lower().replace(" ", "_").replace("-", "_")
-
-                # Check if pack already exists
-                pack_dir = self.soundpacks_dir / pack_id
-                if pack_dir.exists():
-                    # Ask user if they want to overwrite
-                    result = self.app.main_window.question_dialog(
-                        "Pack Already Exists",
-                        f"A sound pack named '{pack_name}' already exists. Do you want to overwrite it?",
-                    )
-                    if not result:
-                        return
-
-                    # Remove existing pack
-                    shutil.rmtree(pack_dir)
-
-                # Extract the sound pack
-                pack_dir.mkdir(exist_ok=True)
-                zip_file.extractall(pack_dir)
-
-                # Reload sound packs
-                self._load_sound_packs()
-                self._refresh_pack_list()
-
-                self.app.main_window.info_dialog(
-                    "Import Successful", f"Sound pack '{pack_name}' has been imported successfully."
-                )
-        except Exception as e:
-            logger.error(f"Failed to import sound pack: {e}")
-            self.app.main_window.error_dialog("Import Error", f"Failed to import sound pack: {e}")
+        _import(self, widget, path)
 
     def _refresh_pack_list(self) -> None:
-        """Refresh the sound pack list."""
-        # Clear existing data
-        self.pack_list.data.clear()
+        from .soundpack_manager.state import refresh_pack_list as _refresh
 
-        # Add pack data to DetailedList
-        for pack_id, pack_info in self.sound_packs.items():
-            pack_name = pack_info.get("name", pack_id)
-            author = pack_info.get("author", "Unknown")
-
-            # Create a data object for the DetailedList
-            pack_data = {
-                "pack_id": pack_id,
-                "pack_info": pack_info,
-                "title": pack_name,
-                "subtitle": f"by {author}",
-                "icon": None,  # Could add icons later
-            }
-
-            self.pack_list.data.append(pack_data)
+        _refresh(self)
 
     def _on_delete_pack(self, widget) -> None:
         """Delete the selected sound pack."""
-        if not self.selected_pack or self.selected_pack == "default":
-            return
+        from .soundpack_manager.ops import delete_pack as _del
 
-        pack_info = self.sound_packs.get(self.selected_pack, {})
-        pack_name = pack_info.get("name", self.selected_pack)
-
-        # Confirm deletion
-        result = self.app.main_window.question_dialog(
-            "Delete Sound Pack",
-            f"Are you sure you want to delete the sound pack '{pack_name}'? This action cannot be undone.",
-        )
-
-        if result:
-            try:
-                pack_path = pack_info.get("path")
-                if pack_path and pack_path.exists():
-                    shutil.rmtree(pack_path)
-
-                # Reload sound packs
-                self._load_sound_packs()
-                self._refresh_pack_list()
-
-                # Clear selection
-                self.selected_pack = None
-                self._update_pack_details()
-                self.select_button.enabled = False
-                self.delete_button.enabled = False
-
-                self.app.main_window.info_dialog(
-                    "Pack Deleted", f"Sound pack '{pack_name}' has been deleted."
-                )
-
-            except Exception as e:
-                logger.error(f"Failed to delete sound pack: {e}")
-                self.app.main_window.error_dialog(
-                    "Delete Error", f"Failed to delete sound pack: {e}"
-                )
+        _del(self, widget)
 
     def _on_create_pack(self, widget) -> None:
-        """Create a new, empty sound pack with default metadata and no sounds.
+        """Create a new, empty sound pack with default metadata and no sounds."""
+        from .soundpack_manager.ops import create_pack as _create
 
-        Generates a unique pack ID (custom_1, custom_2, ...), writes a minimal
-        pack.json, reloads the list, and focuses the new pack.
-        """
-        try:
-            # Find a unique pack ID
-            base = "custom"
-            idx = 1
-            while True:
-                pack_id = f"{base}_{idx}"
-                pack_dir = self.soundpacks_dir / pack_id
-                if not pack_dir.exists():
-                    break
-                idx += 1
-
-            pack_dir.mkdir(parents=True, exist_ok=False)
-            meta = {
-                "name": f"Custom Pack {idx}",
-                "author": "",
-                "description": "A new custom sound pack.",
-                "sounds": {
-                    # Provide common placeholders users can map later
-                    "alert": "alert.wav",
-                    "notify": "notify.wav",
-                },
-            }
-            with open(pack_dir / "pack.json", "w", encoding="utf-8") as f:
-                json.dump(meta, f, indent=2)
-
-            # Refresh in-memory state and UI
-            self._load_sound_packs()
-            self._refresh_pack_list()
-            self.selected_pack = pack_id
-            self.current_pack = pack_id
-            self._update_pack_details()
-
-            if hasattr(self, "duplicate_button"):
-                self.duplicate_button.enabled = True
-            if hasattr(self, "edit_button"):
-                self.edit_button.enabled = True
-            if hasattr(self, "select_button"):
-                self.select_button.enabled = True
-            if hasattr(self, "delete_button"):
-                self.delete_button.enabled = pack_id != "default"
-
-            self.app.main_window.info_dialog(
-                "Pack Created", f"Created new sound pack '{meta['name']}' (ID: {pack_id})."
-            )
-        except Exception as e:
-            logger.error(f"Failed to create sound pack: {e}")
-            self.app.main_window.error_dialog("Create Error", f"Failed to create sound pack: {e}")
+        _create(self)
 
     def _on_duplicate_pack(self, widget) -> None:
         """Duplicate the currently selected pack into a new pack directory."""
-        if not self.selected_pack or self.selected_pack not in self.sound_packs:
-            return
-        try:
-            src_info = self.sound_packs[self.selected_pack]
-            src_dir = src_info.get("path")
-            if not src_dir or not src_dir.exists():
-                return
+        from .soundpack_manager.ops import duplicate_pack as _dup
 
-            # Compute unique destination
-            base = f"{self.selected_pack}_copy"
-            candidate = base
-            suffix = 2
-            while (self.soundpacks_dir / candidate).exists():
-                candidate = f"{base}{suffix}"
-                suffix += 1
-            dst_dir = self.soundpacks_dir / candidate
-
-            shutil.copytree(src_dir, dst_dir)
-
-            # Update metadata name to indicate copy
-            pack_json_path = dst_dir / "pack.json"
-            try:
-                with open(pack_json_path, encoding="utf-8") as f:
-                    meta = json.load(f)
-            except Exception:
-                meta = {}
-            name = meta.get("name", candidate)
-            if "(Copy)" not in name:
-                meta["name"] = f"{name} (Copy)"
-            with open(pack_json_path, "w", encoding="utf-8") as f:
-                json.dump(meta, f, indent=2)
-
-            self._load_sound_packs()
-            self._refresh_pack_list()
-            self.selected_pack = candidate
-            self.current_pack = candidate
-            self._update_pack_details()
-
-            self.app.main_window.info_dialog(
-                "Pack Duplicated",
-                f"Duplicated '{src_info.get('name', self.selected_pack)}' to '{meta.get('name', candidate)}'.",
-            )
-        except Exception as e:
-            logger.error(f"Failed to duplicate sound pack: {e}")
-            self.app.main_window.error_dialog(
-                "Duplicate Error", f"Failed to duplicate sound pack: {e}"
-            )
+        _dup(self)
 
     def _on_edit_pack(self, widget) -> None:
         """Open a small editor to modify name, author, and description of the pack."""
@@ -1019,13 +242,9 @@ class SoundPackManagerDialog:
             return
 
         pack_info = self.sound_packs[self.selected_pack]
-        pack_json_path = pack_info["path"] / "pack.json"
-
         # Load existing metadata
-        meta = {}
         try:
-            with open(pack_json_path, encoding="utf-8") as f:
-                meta = json.load(f)
+            meta = self._load_pack_meta(pack_info)
         except Exception:
             meta = {}
 
@@ -1058,8 +277,7 @@ class SoundPackManagerDialog:
                     "author": (author_input.value or "").strip(),
                     "description": (desc_input.value or "").strip(),
                 }
-                with open(pack_json_path, "w", encoding="utf-8") as f:
-                    json.dump(updated, f, indent=2)
+                self._save_pack_meta(pack_info, updated)
                 # Update in-memory and UI
                 self._load_sound_packs()
                 self._refresh_pack_list()
@@ -1129,73 +347,9 @@ class SoundPackManagerDialog:
             self.app.main_window.error_dialog("Wizard Error", f"Failed to open wizard: {e}")
 
     def _create_pack_from_wizard_state(self, state) -> str:
-        """Create the pack on disk from the wizard state. Returns the new pack_id.
+        from .soundpack_manager.wizard_ops import create_pack_from_wizard_state as _wiz
 
-        This uses similar patterns to _on_create_pack but with user-provided metadata
-        and selected sounds copied from a staging area.
-        """
-
-        # Sanitize and ensure unique ID
-        def _slugify(name: str) -> str:
-            slug = (name or "custom").strip().lower().replace(" ", "_").replace("-", "_")
-            if not slug:
-                slug = "custom"
-            return slug
-
-        base_id = _slugify(getattr(state, "pack_name", ""))
-        if not base_id:
-            base_id = "custom"
-        pack_id = base_id
-        suffix = 1
-        while (self.soundpacks_dir / pack_id).exists():
-            suffix += 1
-            pack_id = f"{base_id}_{suffix}"
-
-        pack_dir = self.soundpacks_dir / pack_id
-        pack_dir.mkdir(parents=True, exist_ok=False)
-
-        # Copy staged files into pack dir and build sounds mapping with filenames
-        sounds: dict[str, str] = {}
-        for key, src in (getattr(state, "sound_mappings", {}) or {}).items():
-            try:
-                src_path = Path(src)
-                if not src_path.exists():
-                    continue
-                dest_name = src_path.name
-                dest_path = pack_dir / dest_name
-                # Avoid collisions: if a different source would overwrite an existing file,
-                # append a numeric suffix to the destination filename.
-                if dest_path.exists():
-                    try:
-                        same = src_path.resolve() == dest_path.resolve()
-                    except Exception:
-                        same = False
-                    if not same:
-                        stem = dest_path.stem
-                        suffix = dest_path.suffix
-                        counter = 1
-                        while True:
-                            candidate = pack_dir / f"{stem}_{counter}{suffix}"
-                            if not candidate.exists():
-                                dest_path = candidate
-                                dest_name = dest_path.name
-                                break
-                            counter += 1
-                if src_path.resolve() != dest_path.resolve():
-                    with contextlib.suppress(Exception):
-                        shutil.copy2(src_path, dest_path)
-                sounds[key] = dest_name
-            except Exception as copy_err:
-                logger.warning(f"Failed to copy sound for {key}: {copy_err}")
-
-        meta = {
-            "name": getattr(state, "pack_name", pack_id) or pack_id,
-            "author": getattr(state, "author", "") or "",
-            "description": getattr(state, "description", "") or "",
-            "sounds": sounds,
-        }
-        with open(pack_dir / "pack.json", "w", encoding="utf-8") as f:
-            json.dump(meta, f, indent=2)
+        return _wiz(self, state)
 
     async def _on_browse_community_packs(self, widget) -> None:
         """Open the community packs browser dialog and refresh after install."""
